begin refactoring of event package into audit package
