+++ audit/formatter_json.go
@@ -1,13 +1,14 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package event
+package audit
 
 import (
 	"context"
 	"fmt"
 
-	vaultaudit "github.com/hashicorp/vault/audit"
+	"github.com/hashicorp/vault/internal/observability/event"
+
 	"github.com/hashicorp/vault/sdk/helper/jsonutil"
 
 	"github.com/hashicorp/eventlogger"
@@ -18,22 +19,20 @@ var _ eventlogger.Node = (*AuditFormatterJSON)(nil)
 // AuditFormatterJSON represents the formatter node which is used to handle
 // formatting audit events as JSON.
 type AuditFormatterJSON struct {
-	config    vaultaudit.FormatterConfig
-	format    auditFormat
-	formatter vaultaudit.Formatter
+	config    FormatterConfig
+	formatter Formatter
 }
 
 // NewAuditFormatterJSON should be used to create an AuditFormatterJSON.
-func NewAuditFormatterJSON(config vaultaudit.FormatterConfig, salter vaultaudit.Salter) (*AuditFormatterJSON, error) {
-	const op = "event.NewAuditFormatterJSON"
+func NewAuditFormatterJSON(config FormatterConfig, salter Salter) (*AuditFormatterJSON, error) {
+	const op = "audit.NewAuditFormatterJSON"
 
-	f, err := vaultaudit.NewAuditFormatter(salter)
+	f, err := NewAuditFormatter(salter)
 	if err != nil {
 		return nil, fmt.Errorf("%s: unable to create new JSON audit formatter: %w", op, err)
 	}
 
 	jsonFormatter := &AuditFormatterJSON{
-		format:    AuditFormatJSON,
 		config:    config,
 		formatter: f,
 	}
@@ -54,7 +53,7 @@ func (_ *AuditFormatterJSON) Type() eventlogger.NodeType {
 // Process will attempt to parse the incoming event data into a corresponding
 // audit request/response entry which is serialized to JSON and stored within the event.
 func (f *AuditFormatterJSON) Process(ctx context.Context, e *eventlogger.Event) (*eventlogger.Event, error) {
-	const op = "event.(AuditFormatterJSON).Process"
+	const op = "audit.(AuditFormatterJSON).Process"
 
 	select {
 	case <-ctx.Done():
@@ -63,18 +62,18 @@ func (f *AuditFormatterJSON) Process(ctx context.Context, e *eventlogger.Event) 
 	}
 
 	if e == nil {
-		return nil, fmt.Errorf("%s: event is nil: %w", op, ErrInvalidParameter)
+		return nil, fmt.Errorf("%s: event is nil: %w", op, event.ErrInvalidParameter)
 	}
 
 	a, ok := e.Payload.(*audit)
 	if !ok {
-		return nil, fmt.Errorf("%s: cannot parse event payload: %w", op, ErrInvalidParameter)
+		return nil, fmt.Errorf("%s: cannot parse event payload: %w", op, event.ErrInvalidParameter)
 	}
 
 	var formatted []byte
 
 	switch a.Subtype {
-	case AuditRequest:
+	case AuditRequestType:
 		entry, err := f.formatter.FormatRequest(ctx, f.config, a.Data)
 		if err != nil {
 			return nil, fmt.Errorf("%s: unable to parse request from audit event: %w", op, err)
@@ -84,7 +83,7 @@ func (f *AuditFormatterJSON) Process(ctx context.Context, e *eventlogger.Event) 
 		if err != nil {
 			return nil, fmt.Errorf("%s: unable to format request: %w", op, err)
 		}
-	case AuditResponse:
+	case AuditResponseType:
 		entry, err := f.formatter.FormatResponse(ctx, f.config, a.Data)
 		if err != nil {
 			return nil, fmt.Errorf("%s: unable to parse response from audit event: %w", op, err)
@@ -98,7 +97,7 @@ func (f *AuditFormatterJSON) Process(ctx context.Context, e *eventlogger.Event) 
 		return nil, fmt.Errorf("%s: unknown audit event subtype: %q", op, a.Subtype)
 	}
 
-	e.FormattedAs(f.format.String(), formatted)
+	e.FormattedAs(AuditFormatJSON.String(), formatted)
 
 	return e, nil
 }
