+++ audit/formatter_jsonx.go
@@ -1,11 +1,13 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package event
+package audit
 
 import (
 	"context"
 	"fmt"
+
+	"github.com/hashicorp/vault/internal/observability/event"
 
 	"github.com/hashicorp/eventlogger"
 	"github.com/jefferai/jsonx"
@@ -14,9 +16,7 @@ import (
 var _ eventlogger.Node = (*AuditFormatterJSONx)(nil)
 
 // AuditFormatterJSONx represents a formatter node which will Process JSON to JSONx format.
-type AuditFormatterJSONx struct {
-	format auditFormat
-}
+type AuditFormatterJSONx struct{}
 
 // NewAuditFormatterJSONx creates a formatter node which can be used to format
 // incoming events to JSONx.
@@ -24,7 +24,7 @@ type AuditFormatterJSONx struct {
 // in the pipeline and will attempt to access the JSON encoded data stored by that
 // formatter node.
 func NewAuditFormatterJSONx() *AuditFormatterJSONx {
-	return &AuditFormatterJSONx{format: AuditFormatJSONx}
+	return &AuditFormatterJSONx{}
 }
 
 // Reopen is a no-op for this formatter node.
@@ -40,7 +40,7 @@ func (_ *AuditFormatterJSONx) Type() eventlogger.NodeType {
 // Process will attempt to retrieve pre-formatted JSON stored within the event
 // and re-encode the data to JSONx.
 func (f *AuditFormatterJSONx) Process(ctx context.Context, e *eventlogger.Event) (*eventlogger.Event, error) {
-	const op = "event.(AuditFormatterJSONx).Process"
+	const op = "audit.(AuditFormatterJSONx).Process"
 
 	select {
 	case <-ctx.Done():
@@ -49,16 +49,16 @@ func (f *AuditFormatterJSONx) Process(ctx context.Context, e *eventlogger.Event)
 	}
 
 	if e == nil {
-		return nil, fmt.Errorf("%s: event is nil: %w", op, ErrInvalidParameter)
+		return nil, fmt.Errorf("%s: event is nil: %w", op, event.ErrInvalidParameter)
 	}
 
 	// We expect that JSON has already been parsed for this event.
 	jsonBytes, ok := e.Format(AuditFormatJSON.String())
 	if !ok {
-		return nil, fmt.Errorf("%s: pre-formatted JSON required but not found: %w", op, ErrInvalidParameter)
+		return nil, fmt.Errorf("%s: pre-formatted JSON required but not found: %w", op, event.ErrInvalidParameter)
 	}
 	if jsonBytes == nil {
-		return nil, fmt.Errorf("%s: pre-formatted JSON required but was nil: %w", op, ErrInvalidParameter)
+		return nil, fmt.Errorf("%s: pre-formatted JSON required but was nil: %w", op, event.ErrInvalidParameter)
 	}
 
 	xmlBytes, err := jsonx.EncodeJSONBytes(jsonBytes)
@@ -69,7 +69,7 @@ func (f *AuditFormatterJSONx) Process(ctx context.Context, e *eventlogger.Event)
 		return nil, fmt.Errorf("%s: encoded JSONx was nil: %w", op, err)
 	}
 
-	e.FormattedAs(f.format.String(), xmlBytes)
+	e.FormattedAs(AuditFormatJSONx.String(), xmlBytes)
 
 	return e, nil
 }
