+++ internal/observability/event/options_test.go
@@ -4,112 +4,11 @@
 package event
 
 import (
-	"os"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/require"
 )
-
-// TestOptions_WithFormat exercises WithFormat option to ensure it performs as expected.
-func TestOptions_WithFormat(t *testing.T) {
-	tests := map[string]struct {
-		Value                string
-		IsErrorExpected      bool
-		ExpectedErrorMessage string
-		ExpectedValue        auditFormat
-	}{
-		"empty": {
-			Value:                "",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "format cannot be empty",
-		},
-		"whitespace": {
-			Value:                "     ",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "format cannot be empty",
-		},
-		"invalid-test": {
-			Value:                "test",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "event.(auditFormat).validate: 'test' is not a valid format: invalid parameter",
-		},
-		"valid-json": {
-			Value:           "json",
-			IsErrorExpected: false,
-			ExpectedValue:   AuditFormatJSON,
-		},
-		"valid-jsonx": {
-			Value:           "jsonx",
-			IsErrorExpected: false,
-			ExpectedValue:   AuditFormatJSONx,
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithFormat(tc.Value)
-			err := applyOption(options)
-			switch {
-			case tc.IsErrorExpected:
-				require.Error(t, err)
-				require.EqualError(t, err, tc.ExpectedErrorMessage)
-			default:
-				require.NoError(t, err)
-				require.Equal(t, tc.ExpectedValue, options.withFormat)
-			}
-		})
-	}
-}
-
-// TestOptions_WithSubtype exercises WithSubtype option to ensure it performs as expected.
-func TestOptions_WithSubtype(t *testing.T) {
-	tests := map[string]struct {
-		Value                string
-		IsErrorExpected      bool
-		ExpectedErrorMessage string
-		ExpectedValue        auditSubtype
-	}{
-		"empty": {
-			Value:                "",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "subtype cannot be empty",
-		},
-		"whitespace": {
-			Value:                "     ",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "subtype cannot be empty",
-		},
-		"valid": {
-			Value:           "AuditResponse",
-			IsErrorExpected: false,
-			ExpectedValue:   AuditResponse,
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithSubtype(tc.Value)
-			err := applyOption(options)
-			switch {
-			case tc.IsErrorExpected:
-				require.Error(t, err)
-				require.EqualError(t, err, tc.ExpectedErrorMessage)
-			default:
-				require.NoError(t, err)
-				require.Equal(t, tc.ExpectedValue, options.withSubtype)
-			}
-		})
-	}
-}
 
 // TestOptions_WithNow exercises WithNow option to ensure it performs as expected.
 func TestOptions_WithNow(t *testing.T) {
@@ -137,7 +36,7 @@ func TestOptions_WithNow(t *testing.T) {
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
 
-			options := &options{}
+			options := &Options{}
 			applyOption := WithNow(tc.Value)
 			err := applyOption(options)
 			switch {
@@ -146,7 +45,7 @@ func TestOptions_WithNow(t *testing.T) {
 				require.EqualError(t, err, tc.ExpectedErrorMessage)
 			default:
 				require.NoError(t, err)
-				require.Equal(t, tc.ExpectedValue, options.withNow)
+				require.Equal(t, tc.ExpectedValue, options.WithNow)
 			}
 		})
 	}
@@ -182,7 +81,7 @@ func TestOptions_WithID(t *testing.T) {
 		tc := tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			options := &options{}
+			options := &Options{}
 			applyOption := WithID(tc.Value)
 			err := applyOption(options)
 			switch {
@@ -191,238 +90,7 @@ func TestOptions_WithID(t *testing.T) {
 				require.EqualError(t, err, tc.ExpectedErrorMessage)
 			default:
 				require.NoError(t, err)
-				require.Equal(t, tc.ExpectedValue, options.withID)
-			}
-		})
-	}
-}
-
-// TestOptions_WithFacility exercises WithFacility option to ensure it performs as expected.
-func TestOptions_WithFacility(t *testing.T) {
-	tests := map[string]struct {
-		Value         string
-		ExpectedValue string
-	}{
-		"empty": {
-			Value:         "",
-			ExpectedValue: "",
-		},
-		"whitespace": {
-			Value:         "    ",
-			ExpectedValue: "",
-		},
-		"value": {
-			Value:         "juan",
-			ExpectedValue: "juan",
-		},
-		"spacey-value": {
-			Value:         "   juan   ",
-			ExpectedValue: "juan",
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithFacility(tc.Value)
-			err := applyOption(options)
-			require.NoError(t, err)
-			require.Equal(t, tc.ExpectedValue, options.withFacility)
-		})
-	}
-}
-
-// TestOptions_WithTag exercises WithTag option to ensure it performs as expected.
-func TestOptions_WithTag(t *testing.T) {
-	tests := map[string]struct {
-		Value         string
-		ExpectedValue string
-	}{
-		"empty": {
-			Value:         "",
-			ExpectedValue: "",
-		},
-		"whitespace": {
-			Value:         "    ",
-			ExpectedValue: "",
-		},
-		"value": {
-			Value:         "juan",
-			ExpectedValue: "juan",
-		},
-		"spacey-value": {
-			Value:         "   juan   ",
-			ExpectedValue: "juan",
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithTag(tc.Value)
-			err := applyOption(options)
-			require.NoError(t, err)
-			require.Equal(t, tc.ExpectedValue, options.withTag)
-		})
-	}
-}
-
-// TestOptions_WithSocketType exercises WithSocketType option to ensure it performs as expected.
-func TestOptions_WithSocketType(t *testing.T) {
-	tests := map[string]struct {
-		Value         string
-		ExpectedValue string
-	}{
-		"empty": {
-			Value:         "",
-			ExpectedValue: "",
-		},
-		"whitespace": {
-			Value:         "    ",
-			ExpectedValue: "",
-		},
-		"value": {
-			Value:         "juan",
-			ExpectedValue: "juan",
-		},
-		"spacey-value": {
-			Value:         "   juan   ",
-			ExpectedValue: "juan",
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithSocketType(tc.Value)
-			err := applyOption(options)
-			require.NoError(t, err)
-			require.Equal(t, tc.ExpectedValue, options.withSocketType)
-		})
-	}
-}
-
-// TestOptions_WithMaxDuration exercises WithMaxDuration option to ensure it performs as expected.
-func TestOptions_WithMaxDuration(t *testing.T) {
-	tests := map[string]struct {
-		Value                string
-		ExpectedValue        time.Duration
-		IsErrorExpected      bool
-		ExpectedErrorMessage string
-	}{
-		"empty-gives-default": {
-			Value: "",
-		},
-		"whitespace-give-default": {
-			Value: "    ",
-		},
-		"bad-value": {
-			Value:                "juan",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "time: invalid duration \"juan\"",
-		},
-		"bad-spacey-value": {
-			Value:                "   juan   ",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "time: invalid duration \"juan\"",
-		},
-		"duration-2s": {
-			Value:         "2s",
-			ExpectedValue: 2 * time.Second,
-		},
-		"duration-2m": {
-			Value:         "2m",
-			ExpectedValue: 2 * time.Minute,
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithMaxDuration(tc.Value)
-			err := applyOption(options)
-			switch {
-			case tc.IsErrorExpected:
-				require.Error(t, err)
-				require.EqualError(t, err, tc.ExpectedErrorMessage)
-			default:
-				require.NoError(t, err)
-				require.Equal(t, tc.ExpectedValue, options.withMaxDuration)
-			}
-		})
-	}
-}
-
-// TestOptions_WithFileMode exercises WithFileMode option to ensure it performs as expected.
-func TestOptions_WithFileMode(t *testing.T) {
-	tests := map[string]struct {
-		Value                string
-		IsErrorExpected      bool
-		ExpectedErrorMessage string
-		IsNilExpected        bool
-		ExpectedValue        os.FileMode
-	}{
-		"empty": {
-			Value:           "",
-			IsErrorExpected: false,
-			IsNilExpected:   true,
-		},
-		"whitespace": {
-			Value:           "     ",
-			IsErrorExpected: false,
-			IsNilExpected:   true,
-		},
-		"nonsense": {
-			Value:                "juan",
-			IsErrorExpected:      true,
-			ExpectedErrorMessage: "unable to parse file mode: strconv.ParseUint: parsing \"juan\": invalid syntax",
-		},
-		"zero": {
-			Value:           "0000",
-			IsErrorExpected: false,
-			ExpectedValue:   os.FileMode(0o000),
-		},
-		"valid": {
-			Value:           "0007",
-			IsErrorExpected: false,
-			ExpectedValue:   os.FileMode(0o007),
-		},
-	}
-
-	for name, tc := range tests {
-		name := name
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			options := &options{}
-			applyOption := WithFileMode(tc.Value)
-			err := applyOption(options)
-			switch {
-			case tc.IsErrorExpected:
-				require.Error(t, err)
-				require.EqualError(t, err, tc.ExpectedErrorMessage)
-			default:
-				require.NoError(t, err)
-				switch {
-				case tc.IsNilExpected:
-					// Optional option 'not supplied' (i.e. was whitespace/empty string)
-					require.Nil(t, options.withFileMode)
-				default:
-					// Dereference the pointer, so we can examine the file mode.
-					require.Equal(t, tc.ExpectedValue, *options.withFileMode)
-				}
+				require.Equal(t, tc.ExpectedValue, options.WithID)
 			}
 		})
 	}
@@ -432,22 +100,17 @@ func TestOptions_WithFileMode(t *testing.T) {
 func TestOptions_Default(t *testing.T) {
 	opts := getDefaultOptions()
 	require.NotNil(t, opts)
-	require.True(t, time.Now().After(opts.withNow))
-	require.False(t, opts.withNow.IsZero())
-	require.Equal(t, "AUTH", opts.withFacility)
-	require.Equal(t, "vault", opts.withTag)
-	require.Equal(t, 2*time.Second, opts.withMaxDuration)
+	require.True(t, time.Now().After(opts.WithNow))
+	require.False(t, opts.WithNow.IsZero())
 }
 
-// TestOptions_Opts exercises getOpts with various Option values.
+// TestOptions_Opts exercises GetOpts with various Option values.
 func TestOptions_Opts(t *testing.T) {
 	tests := map[string]struct {
 		opts                 []Option
 		IsErrorExpected      bool
 		ExpectedErrorMessage string
 		ExpectedID           string
-		ExpectedSubtype      auditSubtype
-		ExpectedFormat       auditFormat
 		IsNowExpected        bool
 		ExpectedNow          time.Time
 	}{
@@ -470,24 +133,24 @@ func TestOptions_Opts(t *testing.T) {
 			ExpectedID:      "juan",
 			IsNowExpected:   true,
 		},
-		"with-multiple-valid-subtype": {
-			opts: []Option{
-				WithSubtype("AuditRequest"),
-				WithSubtype("AuditResponse"),
-			},
-			IsErrorExpected: false,
-			ExpectedSubtype: AuditResponse,
-			IsNowExpected:   true,
-		},
-		"with-multiple-valid-format": {
-			opts: []Option{
-				WithFormat("json"),
-				WithFormat("jsonx"),
-			},
-			IsErrorExpected: false,
-			ExpectedFormat:  AuditFormatJSONx,
-			IsNowExpected:   true,
-		},
+		//"with-multiple-valid-subtype": {
+		//	opts: []Option{
+		//		WithSubtype("AuditRequest"),
+		//		WithSubtype("AuditResponse"),
+		//	},
+		//	IsErrorExpected: false,
+		//	ExpectedSubtype: AuditResponse,
+		//	IsNowExpected:   true,
+		//},
+		//"with-multiple-valid-format": {
+		//	opts: []Option{
+		//		WithFormat("json"),
+		//		WithFormat("jsonx"),
+		//	},
+		//	IsErrorExpected: false,
+		//	ExpectedFormat:  AuditFormatJSONx,
+		//	IsNowExpected:   true,
+		//},
 		"with-multiple-valid-now": {
 			opts: []Option{
 				WithNow(time.Date(2023, time.July, 4, 12, 3, 0, 0, time.Local)),
@@ -508,15 +171,15 @@ func TestOptions_Opts(t *testing.T) {
 		"with-multiple-valid-options": {
 			opts: []Option{
 				WithID("qwerty"),
-				WithSubtype("AuditRequest"),
-				WithFormat("json"),
+				/*	WithSubtype("AuditRequest"),
+					WithFormat("json"),*/
 				WithNow(time.Date(2023, time.July, 4, 12, 3, 0, 0, time.Local)),
 			},
 			IsErrorExpected: false,
 			ExpectedID:      "qwerty",
-			ExpectedSubtype: AuditRequest,
-			ExpectedFormat:  AuditFormatJSON,
-			ExpectedNow:     time.Date(2023, time.July, 4, 12, 3, 0, 0, time.Local),
+			// ExpectedSubtype: AuditRequest,
+			// ExpectedFormat:  AuditFormatJSON,
+			ExpectedNow: time.Date(2023, time.July, 4, 12, 3, 0, 0, time.Local),
 		},
 	}
 
@@ -525,7 +188,7 @@ func TestOptions_Opts(t *testing.T) {
 		tc := tc
 		t.Run(name, func(t *testing.T) {
 			t.Parallel()
-			opts, err := getOpts(tc.opts...)
+			opts, err := GetOpts(tc.opts...)
 
 			switch {
 			case tc.IsErrorExpected:
@@ -534,18 +197,20 @@ func TestOptions_Opts(t *testing.T) {
 			default:
 				require.NotNil(t, opts)
 				require.NoError(t, err)
-				require.Equal(t, tc.ExpectedID, opts.withID)
-				require.Equal(t, tc.ExpectedSubtype, opts.withSubtype)
-				require.Equal(t, tc.ExpectedFormat, opts.withFormat)
+				require.Equal(t, tc.ExpectedID, opts.WithID)
+				// require.Equal(t, tc.ExpectedSubtype, opts.withSubtype)
+				// require.Equal(t, tc.ExpectedFormat, opts.withFormat)
 				switch {
 				case tc.IsNowExpected:
-					require.True(t, time.Now().After(opts.withNow))
-					require.False(t, opts.withNow.IsZero())
+					require.True(t, time.Now().After(opts.WithNow))
+					require.False(t, opts.WithNow.IsZero())
 				default:
-					require.Equal(t, tc.ExpectedNow, opts.withNow)
+					require.Equal(t, tc.ExpectedNow, opts.WithNow)
 				}
 
 			}
 		})
 	}
 }
+
+// TODO: Reproduce some of these tests in audit/options.go
